Release GIL in Tensor indexing functions
Fixes gh-68739

For simple indexing this adds a `gil_scoped_release` before calling
`set_item`. For tuple indexing, the slicing operation is done with the
GIL because otherwise it would have to re-aquire the GIL for each
element in the tuple. However, the GIL is released for the final
`copy_to` operation which is where the actual kernels are called.

ghstack-source-id: b564304ece2cc0ec82e66062ea93343794171093
Pull Request resolved: https://github.com/pytorch/pytorch/pull/71728