dbr quant: rename seen_op_info to seen_q_op_info
Summary:

Renames `seen_op_info` to `seen_q_op_info` and `SeenOpInfo` to `SeenQOpInfo`.

This is to make it clear that the op here is a quantizeable op.
This is useful for a future PR where we will start recording the DAG
of non-quantizeable ops. That will be needed to properly support
function fusion.

Test plan:

CI and mypy

ghstack-source-id: 7df89c685e6cd591f484d6ca321cf826232786d7
Pull Request resolved: https://github.com/pytorch/pytorch/pull/71312