Add OpInfo test to check that floating point inputs in OpInfos have requires_grad set to True
This test detected a number of sampling methods that were not generating
the samples as expected, e.g. `index_put`, `cosine_embedding`, `stft`, but
perhaps most notably the generator for `BinOps`.

It also detected that `reminder` and `fmod` did not have implemented the
backward formula for the second input. I added this in the previous PR.

ghstack-source-id: ca3d51a7b16f6212796e4d52552de57d1d417b12
Pull Request resolved: https://github.com/pytorch/pytorch/pull/69909